# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Survivor Sharks API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: AuthUserRole;
  createdAt?: string;
}

export interface AuthToken {
  token: string;
  user: AuthUser;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  displayName?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type PoolInputSport = typeof PoolInputSport[keyof typeof PoolInputSport];


export const PoolInputSport = {
  nfl: 'nfl',
  mlb: 'mlb',
  nba: 'nba',
  nhl: 'nhl',
  fifa: 'fifa',
} as const;

export interface PoolInput {
  name: string;
  sport: PoolInputSport;
  description?: string;
  maxEntries?: number;
  entryFee?: number;
  prizePot?: number;
  currentWeek?: number;
  season?: number;
}

export interface PoolUpdate {
  name?: string;
  description?: string;
  maxEntries?: number;
  currentWeek?: number;
  season?: number;
  isActive?: boolean;
}

export interface JoinPoolInput {
  inviteCode: string;
}

export interface Pool {
  id: number;
  name: string;
  sport: string;
  /** @nullable */
  description?: string | null;
  inviteCode: string;
  currentWeek: number;
  season: number;
  isActive: boolean;
  memberCount: number;
  activeCount?: number;
  commissionerId: number;
  commissionerName?: string;
  /** @nullable */
  maxEntries?: number | null;
  /** @nullable */
  entryFee?: number | null;
  /** @nullable */
  prizePot?: number | null;
  createdAt?: string;
}

export type PoolMemberStatus = typeof PoolMemberStatus[keyof typeof PoolMemberStatus];


export const PoolMemberStatus = {
  active: 'active',
  eliminated: 'eliminated',
} as const;

export interface PoolMember {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  status: PoolMemberStatus;
  /** @nullable */
  eliminatedWeek?: number | null;
  joinedAt: string;
}

export interface PoolDetail {
  id: number;
  name: string;
  sport: string;
  /** @nullable */
  description?: string | null;
  inviteCode: string;
  currentWeek: number;
  season: number;
  isActive: boolean;
  commissionerId: number;
  commissionerName?: string;
  /** @nullable */
  maxEntries?: number | null;
  /** @nullable */
  entryFee?: number | null;
  /** @nullable */
  prizePot?: number | null;
  members: PoolMember[];
  createdAt?: string;
}

export interface PoolStats {
  poolId: number;
  totalMembers: number;
  activeMembers: number;
  eliminatedMembers: number;
  currentWeek: number;
  weeklyPickRate: number;
  /** @nullable */
  mostPickedTeam: string | null;
  /** @nullable */
  mostPickedTeamCount?: number | null;
  survivorPercentage?: number;
}

export interface PickInput {
  teamId: string;
  week: number;
}

/**
 * @nullable
 */
export type PickResult = typeof PickResult[keyof typeof PickResult] | null;


export const PickResult = {
  win: 'win',
  loss: 'loss',
  pending: 'pending',
} as const;

export interface Pick {
  id: number;
  poolId: number;
  userId: number;
  username?: string;
  teamId: string;
  teamName: string;
  /** @nullable */
  teamLogoUrl?: string | null;
  week: number;
  /** @nullable */
  result: PickResult;
  submittedAt?: string;
}

export interface SurvivorGrid {
  poolId: number;
  weeks: number[];
  members: PoolMember[];
  picks: Pick[];
}

export type LeaderboardEntryStatus = typeof LeaderboardEntryStatus[keyof typeof LeaderboardEntryStatus];


export const LeaderboardEntryStatus = {
  active: 'active',
  eliminated: 'eliminated',
} as const;

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  status: LeaderboardEntryStatus;
  weeksAlive: number;
  /** @nullable */
  eliminatedWeek?: number | null;
  /** @nullable */
  lastPickTeam?: string | null;
  /** @nullable */
  lastPickResult?: string | null;
}

export interface Leaderboard {
  poolId: number;
  currentWeek: number;
  active: LeaderboardEntry[];
  eliminated: LeaderboardEntry[];
}

export interface ProcessResultsInput {
  week: number;
  losingTeamIds: string[];
}

export interface ProcessResultsResult {
  week: number;
  eliminated: string[];
  survived: string[];
  processedAt: string;
}

export interface WeekResult {
  id: number;
  poolId: number;
  week: number;
  losingTeamIds: string[];
  processedAt: string;
  processedBy?: number;
}

export interface Elimination {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  week: number;
  teamId: string;
  teamName: string;
  /** @nullable */
  teamLogoUrl?: string | null;
  eliminatedAt?: string;
}

export interface Team {
  id: string;
  name: string;
  abbreviation: string;
  sport: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  conference?: string | null;
  /** @nullable */
  division?: string | null;
  /** @nullable */
  flagUrl?: string | null;
}

export interface Game {
  id: string;
  sport: string;
  homeTeam: Team;
  awayTeam: Team;
  startTime: string;
  week: number;
  season?: number;
  status?: string;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
}

export type AdminUserRole = typeof AdminUserRole[keyof typeof AdminUserRole];


export const AdminUserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface AdminUser {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: AdminUserRole;
  poolCount?: number;
  createdAt: string;
}

export type AdminUserUpdateRole = typeof AdminUserUpdateRole[keyof typeof AdminUserUpdateRole];


export const AdminUserUpdateRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface AdminUserUpdate {
  role?: AdminUserUpdateRole;
  displayName?: string;
}