# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Survivor Sharks API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Pools ─────────────────────────────────────────────────────────────────
  /pools:
    get:
      operationId: listPools
      tags: [pools]
      summary: List pools the current user belongs to
      responses:
        "200":
          description: List of pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pool"

    post:
      operationId: createPool
      tags: [pools]
      summary: Create a new survivor pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoolInput"
      responses:
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
...
  /pools/{poolId}:
    get:
      operationId: getPool
      tags: [pools]
      summary: Get pool details
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolDetail"
        "404":
          description: Pool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updatePool
      tags: [pools]
      summary: Update pool settings (commissioner only)
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoolUpdate"
      responses:
        "200":
          description: Updated pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"

    delete:
      operationId: deletePool
      tags: [pools]
      summary: Delete a pool (commissioner or admin only)
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Picks ─────────────────────────────────────────────────────────────────
  /pools/{poolId}/picks:
    get:
      operationId: getMyPicks
      tags: [picks]
      summary: Get current user's picks for a pool
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pick"

    post:
      operationId: submitPick
      tags: [picks]
      summary: Submit or update a pick for the current week
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickInput"
      responses:
        "201":
          description: Pick submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pick"
        "400":
          description: Locked or team already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin — update user role or status
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/pools/{poolId}:
    delete:
      operationId: adminDeletePool
      tags: [admin]
      summary: Admin — force delete any pool
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth Schemas ──────────────────────────────────────────────────────────
    AuthToken:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string

    # ── Pool Schemas ──────────────────────────────────────────────────────────
    PoolInput:
      type: object
      required: [name, sport]
      properties:
        name:
          type: string
        sport:
          type: string
          enum: [nfl, mlb, nba, nhl, fifa]
        description:
          type: string
        maxEntries:
          type: integer
        entryFee:
          type: number
        prizePot:
          type: number
        currentWeek:
          type: integer
        season:
          type: integer
    PoolUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        maxEntries:
          type: integer
        currentWeek:
          type: integer
        season:
          type: integer
        isActive:
          type: boolean
...
    Pool:
      type: object
      required: [id, name, sport, inviteCode, currentWeek, season, isActive, memberCount, commissionerId]
      properties:
        id:
          type: integer
        name:
          type: string
        sport:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        currentWeek:
          type: integer
        season:
          type: integer
        isActive:
          type: boolean
        memberCount:
          type: integer
        activeCount:
          type: integer
        commissionerId:
          type: integer
        commissionerName:
          type: string
        maxEntries:
          type: ["integer", "null"]
        entryFee:
          type: ["number", "null"]
        prizePot:
          type: ["number", "null"]
        createdAt:
          type: string
    PoolDetail:
      type: object
      required: [id, name, sport, inviteCode, currentWeek, season, isActive, members, commissionerId]
      properties:
        id:
          type: integer
        name:
          type: string
        sport:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        currentWeek:
          type: integer
        season:
          type: integer
        isActive:
          type: boolean
        commissionerId:
          type: integer
        commissionerName:
          type: string
        maxEntries:
          type: ["integer", "null"]
        entryFee:
          type: ["number", "null"]
        prizePot:
          type: ["number", "null"]
        members:
          type: array
          items:
            $ref: "#/components/schemas/PoolMember"
        createdAt:
          type: string
    PoolMember:
      type: object
      required: [userId, username, status, joinedAt]
      properties:
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, eliminated]
        eliminatedWeek:
          type: ["integer", "null"]
        joinedAt:
          type: string
...
    PickInput:
      type: object
      required: [teamId, week]
      properties:
        teamId:
          type: string
        week:
          type: integer
    Pick:
      type: object
      required: [id, poolId, userId, teamId, teamName, week, result]
      properties:
        id:
          type: integer
        poolId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        teamId:
          type: string
        teamName:
          type: string
        teamLogoUrl:
          type: ["string", "null"]
        week:
          type: integer
        result:
          type: ["string", "null"]
          enum: [win, loss, pending, null]
        submittedAt:
          type: string
...
    AdminUser:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        poolCount:
          type: integer
        createdAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        displayName:
          type: string