# lib/api-zod/src/generated/types/authToken.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Survivor Sharks API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthToken {
  token: string;
  user: AuthUser;
}